pyproject.toml
[project]
name = "stt2desktop"
dynamic = ["version"]
description = "Local speech-to-text for desktop using faster-whisper"
license = {text = "GPL-3.0-or-later"}
readme = "README.md"
authors = [
    {name = 'Jens Diemer', email = 'stt2desktop@jensdiemer.de'}
]
requires-python = ">=3.12"
dependencies = [
    "faster-whisper",  # https://github.com/SYSTRAN/faster-whisper
    "sounddevice",  # https://github.com/spatialaudio/python-sounddevice/
    "pynput",  # https://github.com/moses-palmer/pynput
    "cli-base-utilities>=0.27.1",  # https://github.com/jedie/cli-base-utilities
    "bx_py_utils",  # https://github.com/boxine/bx_py_utils
    "tyro",  # https://github.com/brentyi/tyro
    "rich",  # https://github.com/Textualize/rich
]

[dependency-groups]
dev = [
    "manageprojects",  # https://github.com/jedie/manageprojects
    "urllib3", # for bx_py_utils.test_utils.deny_requests.deny_any_real_request() in tests
    "uv",  # https://github.com/astral-sh/uv
    "hatchling",  # https://github.com/pypa/hatch/tree/master/backend
    "nox",  # https://github.com/wntrblm/nox
    "coverage",  # https://github.com/nedbat/coveragepy
    "ruff", # https://github.com/astral-sh/ruff
    "codespell",  # https://github.com/codespell-project/codespell
    "EditorConfig",  # https://github.com/editorconfig/editorconfig-core-py
    "pip-audit",  # https://github.com/pypa/pip-audit
    "mypy",  # https://github.com/python/mypy
    "twine",  # https://github.com/pypa/twine
    "pre-commit",  # https://github.com/pre-commit/pre-commit
    "typeguard",  # https://github.com/agronholm/typeguard/
]


[tool.uv]
# Don't install bleed-edge versions of dependencies, to avoid undetected issues and supply chain attacks.
exclude-newer = "1 week"
# Exclude external packages temporarily to fix known issues or current CVEs
# TODO: Remove requests - v2.33.0 fixed CVE-2026-25645
exclude-newer-package = { requests = "2026-03-25T18:00:00Z" }


[project.urls]
Documentation = "https://github.com/jedie/stt2desktop"
Source = "https://github.com/jedie/stt2desktop"

[project.scripts]
stt2desktop = "stt2desktop.__main__:main"


[tool.bx_py_utils.doc_write]
docstring_prefix = 'DocWrite:'
macro_prefix = 'DocWriteMacro:'
output_base_path = './docs/'
search_paths = ['./stt2desktop/']
delete_obsolete_files = true


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["stt2desktop"]

[tool.hatch.version]
path = "stt2desktop/__init__.py"


[tool.cli_base.pip_audit]
strict=true
require_hashes=true
ignore-vuln=[
    "CVE-2026-4539", # https://github.com/pygments/pygments/issues/3058#issuecomment-4127344972
]

[tool.cli_base]
version_module_name = "stt2desktop"  # Used by "update-readme-history" git hook


[tool.ruff]
# https://docs.astral.sh/ruff/configuration/
line-length = 120
exclude = [
  ".*/",
]

[tool.ruff.lint]
preview = true  # Needed for some of the rules
extend-select = [
    "E", # pycodestyle: https://docs.astral.sh/ruff/rules/#error-e
    "I", # isort: https://docs.astral.sh/ruff/rules/#isort-i
]
ignore = [
    "E402", # https://docs.astral.sh/ruff/rules/module-import-not-at-top-of-file/
    "E731", # https://docs.astral.sh/ruff/rules/lambda-assignment/
    "PLW1508", # Invalid type for environment variable default
]

[tool.ruff.lint.isort]
# https://docs.astral.sh/ruff/settings/#lintisort
lines-after-imports = 2

[tool.ruff.format]
quote-style = "single"


[tool.coverage.run]
branch = true
parallel = true
concurrency = ["multiprocessing"]
source = ['.']
command_line = '-m unittest --verbose --locals --buffer'

[tool.coverage.report]
omit = ['.*', '*/tests/*']
skip_empty = true
fail_under = 30
show_missing = true
exclude_lines = [
    'if self.debug:',
    'pragma: no cover',
    'raise NotImplementedError',
    'if __name__ == .__main__.:',
]


[tool.mypy]
warn_unused_configs = true
ignore_missing_imports = true
allow_redefinition = true  # https://github.com/python/mypy/issues/7165
show_error_codes = true
plugins = []
exclude = ['.venv', 'tests']


[manageprojects] # https://github.com/jedie/manageprojects
initial_revision = "af6dd62"
initial_date = 2026-01-26T16:04:53+01:00
cookiecutter_template = "https://github.com/jedie/cookiecutter_templates/"
cookiecutter_directory = "uv-python"
applied_migrations = [
    "af6dd62", # 2026-01-26T16:04:53+01:00
]

[manageprojects.cookiecutter_context.cookiecutter]
full_name = "Jens Diemer"
github_username = "jedie"
author_email = "stt2desktop@jensdiemer.de"
package_name = "stt2desktop"
package_version = "0.0.1"
package_description = "local STT for KDE using faster-whisper"
package_url = "https://github.com/jedie/stt2desktop"
issues_url = "https://github.com/jedie/stt2desktop/issues"
license = "GPL-3.0-or-later"
_template = "https://github.com/jedie/cookiecutter_templates/"
